bitcoin_indexer::indexer
Struct MockIndexer

pub struct MockIndexer<B>
where
    B: BitcoinClientApi,
{ /* private fields */ }

Implementations

impl<B> MockIndexer<B>
where B: BitcoinClientApi,

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl<B> MockIndexer<B>
where B: BitcoinClientApi,

pub fn expect_is_ready(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the is_ready method

pub fn expect_get_best_height(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_best_height method

pub fn expect_get_blockchain_best_height(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_blockchain_best_height method

pub fn expect_get_best_block(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_best_block method

pub fn expect_get_tx(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_tx method

pub fn expect_get_estimated_fee_rate(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_estimated_fee_rate method

pub fn expect_tick(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the tick method

pub fn expect_get_block_by_height(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_block_by_height method

pub fn expect_get_block_by_hash(&mut self) -> &mut Expectation<B>

Create an Expectation for mocking the get_block_by_hash method