bitcoin_indexer::indexer
Trait IndexerApi

pub trait IndexerApi {
    // Required methods
    fn is_ready(&self) -> Result<bool, IndexerError>;
    fn tick(&self) -> Result<(), IndexerError>;
    fn get_best_block(&self) -> Result<Option<FullBlock>, IndexerError>;
    fn get_best_height(&self) -> Result<Option<BlockHeight>, IndexerError>;
    fn get_blockchain_best_height(&self) -> Result<BlockHeight, IndexerError>;
    fn get_block_by_height(
        &self,
        height: BlockHeight,
    ) -> Result<Option<FullBlock>, IndexerError>;
    fn get_block_by_hash(
        &self,
        hash: &BlockHash,
    ) -> Result<Option<FullBlock>, IndexerError>;
    fn get_tx(
        &self,
        tx_id: &Txid,
    ) -> Result<Option<TransactionInfo>, IndexerError>;
    fn get_estimated_fee_rate(&self) -> Result<u64, IndexerError>;
}

Required Methods

fn is_ready(&self) -> Result<bool, IndexerError>

Checks if the indexer has indexed the entire blockchain and is at the latest block. Returns Ok(true) if ready, Ok(false) if not, or an IndexerError if an error occurs.

fn tick(&self) -> Result<(), IndexerError>

Processes the next block in the blockchain. Returns Ok(()) if successful, or an IndexerError if an error occurs during processing.

fn get_best_block(&self) -> Result<Option<FullBlock>, IndexerError>

Retrieves the best (most recent) block that has been indexed. Returns Ok(Some(FullBlock)) if a block is found, Ok(None) if no block was indexed, or an IndexerError if an error occurs.

fn get_best_height(&self) -> Result<Option<BlockHeight>, IndexerError>

Retrieves the height of the best (most recent) block that has been indexed. Returns Ok(Some(BlockHeight)) if a height is found, Ok(None) if no block is indexed, or an IndexerError if an error occurs.

fn get_blockchain_best_height(&self) -> Result<BlockHeight, IndexerError>

Retrieves the current best block height from the Bitcoin blockchain. Returns Ok(BlockHeight) with the current blockchain height, or an IndexerError if an error occurs.

fn get_block_by_height( &self, height: BlockHeight, ) -> Result<Option<FullBlock>, IndexerError>

Retrieves a block by its height. Returns Ok(Some(FullBlock)) if the block is found, Ok(None) if not found, or an IndexerError if an error occurs.

fn get_block_by_hash( &self, hash: &BlockHash, ) -> Result<Option<FullBlock>, IndexerError>

fn get_tx(&self, tx_id: &Txid) -> Result<Option<TransactionInfo>, IndexerError>

Retrieves transaction information for a given transaction ID. Returns Ok(Some(TransactionInfo)) if the transaction is found, Ok(None) if not found, or an IndexerError if an error occurs.

fn get_estimated_fee_rate(&self) -> Result<u64, IndexerError>

Retrieves the estimated fee rate from the most recently indexed block. Returns Ok(u64) with the fee rate in satoshis per virtual byte (sat/vB), or an IndexerError if an error occurs or the indexer is not synced.