bitcoin_indexer::store
Struct MockIndexerStore

pub struct MockIndexerStore { /* private fields */ }

Implementations

impl MockIndexerStore

pub fn checkpoint(&mut self)

Validate that all current expectations for all methods have been satisfied, and discard them.

pub fn new() -> Self

Create a new mock object with no expectations.

This method will not be generated if the real struct already has a new method. However, it will be generated if the struct implements a trait with a new method. The trait’s new method can still be called like <MockX as TraitY>::new

impl MockIndexerStore

pub fn expect_save_new_best_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_new_best_block method

pub fn expect_get_best_block(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_best_block method

pub fn expect_get_block_hash_by_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_block_hash_by_height method

pub fn expect_get_block_by_hash(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_block_by_hash method

pub fn expect_get_tx_info(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_tx_info method

pub fn expect_get_block_by_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_block_by_height method

pub fn expect_get_best_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_best_height method

pub fn expect_save_best_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_best_height method

pub fn expect_mark_following_blocks_as_orphan(&mut self) -> &mut Expectation

Create an Expectation for mocking the mark_following_blocks_as_orphan method

pub fn expect_get_checkpoint_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the get_checkpoint_height method

pub fn expect_save_checkpoint_height(&mut self) -> &mut Expectation

Create an Expectation for mocking the save_checkpoint_height method